refactor: AppSection component
