+++ src/components/AppSection/index.js
@@ -1,0 +1,3 @@
+import AppSection from './AppSection';
+
+export { AppSection as default, AppSection };
